requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
pandas>=2.1.0
openpyxl>=3.1.2
google-genai>=0.3.0
gspread>=6.0.0
google-auth>=2.27.0
PyPDF2>=3.0.1
python-docx>=1.1.0
